LayoutTests/canvas/philip/tests/2d.imageData.get.source.negative.html

<!DOCTYPE html>
<title>Canvas test: 2d.imageData.get.source.negative</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<link rel="prev" href="2d.imageData.get.source.outside.html" title="2d.imageData.get.source.outside">
<link rel="next" href="2d.imageData.get.source.size.html" title="2d.imageData.get.source.size">
<body class="show_output">
<p>
 <a href="2d.imageData.get.source.outside.html" accesskey="p" title="[p] 2d.imageData.get.source.outside">&lt;</a>
 <a href="index.html">[index]</a>
 <a href="2d.imageData.get.source.size.html" accesskey="n" title="[n] 2d.imageData.get.source.size">&gt;</a>
<h1><a href="index.2d.html">2d</a>.<a href="index.2d.imageData.html">imageData</a>.<a href="index.2d.imageData.get.html">get</a>.<a href="index.2d.imageData.get.source.html">source</a>.negative</h1>
<p class="desc">getImageData() works with negative width and height, and returns top-to-bottom left-to-right</p>
<div class="refs">References:
<ul>
<li><a href="spec.html#testrefs.2d.imageData.get.basic">2d.imageData.get.basic</a>
<li><a href="spec.html#testrefs.2d.pixelarray.order">2d.pixelarray.order</a>

</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#000';
ctx.fillRect(0, 0, 100, 50);
ctx.fillStyle = '#fff';
ctx.fillRect(20, 10, 60, 10);

var imgdata1 = ctx.getImageData(85, 25, -10, -10);
_assertSame(imgdata1.data[0], 255, "imgdata1.data[\""+(0)+"\"]", "255");
_assertSame(imgdata1.data[1], 255, "imgdata1.data[\""+(1)+"\"]", "255");
_assertSame(imgdata1.data[2], 255, "imgdata1.data[\""+(2)+"\"]", "255");
_assertSame(imgdata1.data[3], 255, "imgdata1.data[\""+(3)+"\"]", "255");
_assertSame(imgdata1.data[imgdata1.data.length-4+0], 0, "imgdata1.data[imgdata1.data.length-4+0]", "0");
_assertSame(imgdata1.data[imgdata1.data.length-4+1], 0, "imgdata1.data[imgdata1.data.length-4+1]", "0");
_assertSame(imgdata1.data[imgdata1.data.length-4+2], 0, "imgdata1.data[imgdata1.data.length-4+2]", "0");
_assertSame(imgdata1.data[imgdata1.data.length-4+3], 255, "imgdata1.data[imgdata1.data.length-4+3]", "255");

var imgdata2 = ctx.getImageData(0, 0, -1, -1);
_assertSame(imgdata2.data[0], 0, "imgdata2.data[\""+(0)+"\"]", "0");
_assertSame(imgdata2.data[1], 0, "imgdata2.data[\""+(1)+"\"]", "0");
_assertSame(imgdata2.data[2], 0, "imgdata2.data[\""+(2)+"\"]", "0");
_assertSame(imgdata2.data[3], 0, "imgdata2.data[\""+(3)+"\"]", "0");


});
</script>
